pyproject.toml
[tool.poetry]
name = "omnia_timeseries"
version = "1.4.0"
authors = ["Equinor Omnia Industrial IoT Team <omniaindiot@equinor.com>"]
homepage = "https://github.com/equinor/omnia-timeseries-python"
repository = "https://github.com/equinor/omnia-timeseries-python"
description = "Official Python SDK for the Omnia Timeseries API"
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "*"
azure-identity = "^1.10.0"
azure-core = "*"
msal-bearer = "^0.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
requests-mock = "^1.11.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
opentelemetry-instrumentation-requests = "^0.41b0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"